+++ packages/kbn-test/src/mocha/__tests__/junit_report_generation.js
