+++ packages/kbn-test/src/mocha/auto_junit_reporter.js
