+++ packages/kbn-test/src/mocha/index.ts
@@ -23,3 +23,7 @@ export { createAutoJUnitReporter } from './auto_junit_reporter';
 export { setupJUnitReportGeneration } from './junit_report_generation';
 // @ts-ignore not typed yet
 export { runMochaCli } from './run_mocha_cli';
+// @ts-ignore not typed yet
+export { recordLog, snapshotLogsForRunnable } from './log_cache';
+// @ts-ignore not typed yet
+export { escapeCdata } from './xml';
