+++ packages/kbn-test/src/mocha/junit_report_generation.js
@@ -25,14 +25,14 @@ import xmlBuilder from 'xmlbuilder';
 import { getUniqueJunitReportPath } from '@kbn/test';
 
 import { getSnapshotOfRunnableLogs } from './log_cache';
-import { escapeCdata } from '../xml';
+import { escapeCdata } from '../';
 
 const dateNow = Date.now.bind(Date);
 
 export function setupJUnitReportGeneration(runner, options = {}) {
   const {
     reportName = 'Unnamed Mocha Tests',
-    rootDirectory = dirname(require.resolve('../../../package.json')),
+    rootDirectory = dirname(require.resolve('../../../../package.json')),
     getTestMetadata = () => ({}),
   } = options;
