+++ packages/kbn-test/src/mocha/run_mocha_cli.js
@@ -40,9 +40,6 @@ export function runMochaCli() {
   // prevent globals injected from canvas plugins from triggering leak check
   process.argv.push('--globals', '__core-js_shared__,core,_, ');
 
-  // ensure that mocha requires the setup_node_env script
-  process.argv.push('--require', require.resolve('../../setup_node_env'));
-
   // set default test timeout
   if (opts.timeout == null && !opts['no-timeouts']) {
     if (runInBand) {
