+++ scripts/mocha.js
@@ -18,4 +18,4 @@
  */
 
 require('../src/setup_node_env');
-require('../src/dev/mocha').runMochaCli();
+require('@kbn/test').runMochaCli();
