+++ src/dev/index.js
@@ -17,6 +17,6 @@
  * under the License.
  */
 
-export { createAutoJUnitReporter, setupJUnitReportGeneration } from './mocha';
+export { createAutoJUnitReporter, setupJUnitReportGeneration } from '@kbn/test';
 
 export { generateNoticeFromSource } from './notice';
